Functions

Index

AbstractItemResponseModels.expected_score

AbstractItemResponseModels.iif

AbstractItemResponseModels.information

AbstractItemResponseModels.irf

StatsAPI.fit

Model fitting

StatsAPI.fit — Function

fit(modeltype::Type{<: AbstractRaschModel}, data, alg, args...; kwargs...)

Fit a Rasch Model to response data.

Arguments

modeltype

An AbstractRaschModel type. Currently RaschModels.jl implements the following model types:

PartialCreditModel

RaschModel

RatingScaleModel

data

The observed data matrix. It is assumed that each person corresponds to a row and each item to a column in the matrix.

For dichotomous models correct responses are coded as 1 and incorrect responses are coded as 0.

For polytomous models, categories are coded as an integer sequence starting from 1. If your model features e.g. three categories, they are coded as 1, 2 and 3 respectively.

alg

The estimation algorithm.

The following algorithm is still in development phase, by now only the RaschModel is supported.

CML: Conditional maximum likelihood

For (bayesian) sampling based estimation all Turing.InferenceAlgorithm types are supported. The following algorithms are reexported from this package:

NUTS: No-U-Turn-Sampler

HMC: Hamiltonian Monte Carlo

MH: Metropolis-Hastings

Bayesian point estimation is supported by

MAP: Maximum a posteriori (Posterior mode)

MLE: Maximum Likelihood (Posterior mean)

Examples

Conditional maximum likelihood estimation

data = rand(0:1, 100, 10)
rasch_fit = fit(RaschModel, data, CML())

Bayesian estimation

Fitting a simple Rasch model with the No-U-Turn-Sampler. Note that estimation with Turing.InterenceAlgorithm uses the AbstractMCMC.sample interface and thus requires an additional argument to specify the number of iterations.

data = rand(0:1, 100, 10)
rasch_fit = fit(RaschModel, data, NUTS(), 1_000)

You can also sample multiple chains in parallel.

data = rand(0:1, 100, 10)
rasch_fit = fit(RaschModel, data, NUTS(), MCMCThreads(), 1_000, 4)

For bayesian point estimation you can just swap the inference algorithm to one of the supported point estimation algorithms.

data = rand(0:1, 100, 10)
rasch_fit = fit(RaschModel, data, MAP())

Model evaluation

AbstractItemResponseModels.irf — Function

irf(model::RaschModel, theta, i, y)
irf(model::RaschModel, theta, i)

Evaluate the item response function for a dichotomous Rasch model for item i at the ability value theta.

If the response value y is omitted, the item response probability for a correct response y = 1 is returned.

Examples

Point estimation

julia> data = rand(0:1, 50, 3);

julia> rasch = fit(RaschModel, data, MLE());

julia> irf(rasch, 0.0, 1)
0.3989070983504997

Bayesian estimation

julia> data = rand(0:1, 50, 3);

julia> rasch = fit(RaschModel, data, MH(), 100; progress = false);

julia> irf(rasch, 0.0, 1);

irf(model::PolytomousRaschModel, theta, i, y)
irf(model::PolytomousRaschModel, theta, i)

Evaluate the item response function for a polytomous Rasch model (Partial Credit Model or Rating Scale Model) for item i at the ability value theta.

If the response value y is omitted, the item response probabilities for each category are returned. To calculate expected scores for an item, see expected_score.

AbstractItemResponseModels.iif — Function

iif(model::RaschModel, theta, i, y)
iif(model::RaschModel, theta, i)

Evaluate the item information function for a dichotomous Rasch model for item i at the ability value theta.

If the response value y is omitted, the item information for a correct response y = 1 is returned.

Examples

Point estimation

julia> data = rand(0:1, 50, 3);

julia> rasch = fit(RaschModel, data, MLE());

julia> iif(rasch, 0.0, 1)
0.07287100351275909

Bayesian estimation

julia> data = rand(0:1, 50, 3);

julia> rasch = fit(RaschModel, data, MH(), 100; progress = false);

julia> iif(rasch, 0.0, 1);

iif(model::PolytomousRaschModel, theta, i, y)
iif(model::PolytomousRaschModel, theta, i)

Evaluate the item (category) information function for a polytomous Rasch model (Partial Credit Model or Rating Scale Model) for item i at the ability value theta.

If the response value y is omitted, the item information for each category is returned. To calculate the total information of an item, see @ref.

AbstractItemResponseModels.expected_score — Function

expected_score(model::RaschModel, theta, is; scoring_function)
expected_score(model::RaschModel, theta; scoring_function)

Calculate the expected score for a dichotomous Rasch model at ability value theta for a set of items is.

is can either be a single item index, an array of item indices, or a range of values. If is is omitted, the expected score for the whole test is calculated.

scoring_function can be used to add weights to the resulting expected scores (see Examples for details).

Examples

Point estimation

julia> data = rand(0:1, 50, 3);

julia> rasch = fit(RaschModel, data, MLE());

julia> expected_score(rasch, 0.0)  # all 3 items
0.4435592483649984

julia> expected_score(rasch, 0.0, 1:2)  # items 1 and 2
0.3054954277378461

julia> expected_score(rasch, 0.0, [1, 3])  # items 1 and 3
0.21719686244768088

Bayesian estimation

julia> data = rand(0:1, 50, 3);

julia> rasch = fit(RaschModel, data, MH(), 100; progress = false);

julia> expected_score(rasch, 0.0);  # all 3 items

julia> expected_score(rasch, 0.0, 1:2);  # items 1 and 2

julia> expected_score(rasch, 0.0, [1, 3]);  # items 1 and 3

Using the scoring function

Using the scoring_function keyword argument allows to weigh response probabilities by a value depending on the response y. It is of the form f(y) = x, assigning a scalar value to every possible reponse value y.

For the Rasch Model the valid responses are 0 and 1. If we want to calculate expected scores doubling the weight for y = 1, the weighted responses are 0 and 2. The corresponding scoring_function is y -> 2y,

julia> data = rand(0:1, 50, 3);

julia> rasch = fit(RaschModel, data, MLE());

julia> expected_score(rasch, 0.0; scoring_function = y -> 2y)
4.592642952772852

expected_score(model::PolytomousRaschModel, theta, is)
expected_score(model::PolytomousRaschModel, theta)

Calculate the expected score for a polytomous Rasch model at theta for a set of items is.

is can either be a single item index, an array of item indices, or a range of values. If is is omitted, the expected score for the whole test is calculated.

AbstractItemResponseModels.information — Function

information(model::RaschModel, theta, is; scoring_function)
information(model::RaschModel, theta; scoring_function)

Calculate the information for a dichotomous Rasch model at the ability value theta for a set of items is.

is can either be a single item index, an array of item indices, or a range of values. If is is omitted, the information for the whole test is calculated.

scoring_function can be used to add weights to the resulting information (see Examples for details).

Examples

Point estimation

julia> data = rand(0:1, 50, 3);

julia> rasch = fit(RaschModel, data, MLE());

julia> information(rasch, 0.0)  # all 3 items
0.519893299555712

julia> information(rasch, 0.0, 1:2)  # items 1 and 2
0.4121629113381827

julia> information(rasch, 0.0, [1, 3])  # items 1 and 3
0.313811843802304

Bayesian estimation

julia> data = rand(0:1, 50, 3);

julia> rasch = fit(RaschModel, data, MH(), 100; progress = false);

julia> information(rasch, 0.0);  # all 3 items

julia> information(rasch, 0.0, 1:2);  # items 1 and 2

julia> information(rasch, 0.0, [1, 3]);  # items 1 and 3

Using the scoring function

Using the scoring_function keyword argument allows to weigh response probabilities by a value depending on the response y. It is of the form f(y) = x, assigning a scalar value to every possible reponse value y.

For the Rasch Model the valid responses are 0 and 1. If we want to calculate the information doubling the weight for y = 1, the weighted responses are 0 and 2. The corresponding scoring_function is y -> 2y,

julia> data = rand(0:1, 50, 3);

julia> rasch = fit(RaschModel, data, MLE());

julia> information(rasch, 0.0; scoring_function = y -> 2y)
2.079573198222848

information(model::PolytomousRaschModel, theta, is)
information(model::PolytomousRaschModel, theta)

Calculate the information for a polytomous Rasch model at theta for a set of items is.

is can either be a single item index, an array of item indices, or a range of values. If is is omitted, the information for the whole test is calculated.